pyproject.toml
[tool.poetry]
name = "tmnt-bot"
version = "0.1.0"
description = ""
authors = ["Sax Dhiut <mail@iyxeyl.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
spacy = "^3.7.2"
en_core_web_sm = {url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1.tar.gz"}
ru_core_news_sm = {url = "https://github.com/explosion/spacy-models/releases/download/ru_core_news_sm-3.7.0/ru_core_news_sm-3.7.0.tar.gz"}
pymorphy3 = "^1.3.1"
pyphen = "^0.14.0"
lingua-language-detector = "^2.0.2"
aiogram = "^3.2.0"
pydantic-settings = "^2.1.0"
pydantic = "^2.5.2"
result = "^0.15.0"

[tool.poetry.group.dev.dependencies]
black = "^23.12.0"
mypy = "^1.7.1"
isort = "^5.13.2"
ruff = "^0.1.8"
vulture = "^2.10"


[tool.isort]
profile = "black"

[tool.mypy]
pretty = true
check_untyped_defs = true
disallow_incomplete_defs = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_defs = true
warn_unreachable = true


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"